Add routes and controller for updating category .Enhance operations by adding validation with proper error messages.  Modify category schema to ensure name is stored in lowercase.

+++ backend/src/controllers/category.controllers.js
@@ -2,8 +2,11 @@ import Category from "../models/category.model.js";
 import ApiError from "../utils/ApiError.js";
 import ApiResponse from "../utils/ApiResponse.js";
 import asyncHandler from "../utils/asyncHandler.js";
+import { validationResult } from "express-validator";
 
 export const createCategory = asyncHandler(async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) throw ApiError.validationError(result.array());
   const { name, description, isActive } = req.body;
 
   const alreadyExists = await Category.findOne({ name });
@@ -22,13 +25,14 @@ export const createCategory = asyncHandler(async (req, res) => {
 });
 
 export const getAllCategories = asyncHandler(async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) throw ApiError.validationError(result.array());
+
   const { limit = 10, page = 1 } = req.query;
   const categories = await Category.aggregate([
     {
-      $project: {
-        name: 1,
-        description: 1,
-        isActive: 1,
+      $sort: {
+        createdAt: -1,
       },
     },
     {
@@ -36,6 +40,13 @@ export const getAllCategories = asyncHandler(async (req, res) => {
     },
     {
       $limit: Number(limit),
+    },
+    {
+      $project: {
+        name: 1,
+        description: 1,
+        isActive: 1,
+      },
     },
   ]);
 
@@ -54,7 +65,39 @@ export const getAllCategories = asyncHandler(async (req, res) => {
     .json(ApiResponse.success(dataToSend, "category fetched successfully"));
 });
 
+export const updateCategory = asyncHandler(async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) throw ApiError.validationError(result.array());
+
+  const { id } = req.params;
+  const { name, description, isActive } = req.body;
+
+  const category = await Category.findById(id);
+  if (!category) throw new ApiError(400, "category not found");
+  console.log(req.body);
+
+  const updatedCategory = await Category.findByIdAndUpdate(
+    id,
+    {
+      name,
+      description,
+      isActive,
+    },
+    { new: true }
+  );
+  if (!updatedCategory) throw new ApiError(500, "Failed to update category");
+
+  return res
+    .status(200)
+    .json(
+      ApiResponse.success(updatedCategory, "Category updated successfully!")
+    );
+});
+
 export const deleteCategoryById = asyncHandler(async (req, res) => {
+  const result = validationResult(req);
+  if (!result.isEmpty()) throw ApiError.validationError(result.array());
+
   const { id } = req.params;
 
   const category = await Category.findByIdAndDelete(id);
